Functionally working now with serial2ws example code
A problem is that it seems like all the methods in PrimareTalker.setup()
is called first before we start listening, so for some reason model and
input is not being read (maybe the buffer lacks room?
Doesn't seem like the cause as we ask for 15, 16, 17, 14 - and get
replies back for 15 and 17.

Need to implement and test the WAMP RPC and pub/sub also